Add files via upload
Part one is used to do the image processing steps and it creates a table of values stored as an excel sheet.

Part two uses the excel sheet and implements the data in machine learning algorithms. It requires the table titled 'Summary.xls' that is provided in the repository. Also the file
uses the function 'algorithmerror' and so that also has to be in the same folder as the program itself.